Fix token expiration check in useSession hook.

+++ src/hooks/useSession.js
@@ -10,12 +10,14 @@ const useSession = () => {
   const navigate = useNavigate();
 
   const checkTokenExpirationTime = () => {
-    const convertUnixDateToMillisecond = decodedSession.exp * 1000;
-    const expirationDate = new Date(convertUnixDateToMillisecond);
-    const currentDate = new Date();
+    if (decodedSession && decodedSession.exp) {
+      const convertUnixDateToMillisecond = decodedSession.exp * 1000;
+      const expirationDate = new Date(convertUnixDateToMillisecond);
+      const currentDate = new Date();
 
-    if (expirationDate < currentDate) {
-      localStorage.clear();
+      if (expirationDate < currentDate) {
+        localStorage.clear();
+      }
     }
   };
